requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
chamfer==2.0.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cubic_feature_sampling==1.1.0
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
einops==0.8.0
emd==0.0.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
future==1.0.0
gpustat==1.1.1
gridding==2.1.0
gridding_distance==1.0.0
grpcio==1.65.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.35.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
kornia==0.6.0
lightning-lite==1.8.0
lightning-utilities==0.3.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py==12.555.43
omegaconf==2.3.0
open-clip-torch==2.0.2
open3d==0.18.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
plyfile==1.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyDeprecate==0.3.1
pyglet==2.0.17
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyre-extensions==0.0.23
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==1.8.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
tenacity==9.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.7
tinycss2==1.3.0
tokenizers==0.12.1
tomli==2.0.1
torch @ https://download.pytorch.org/whl/cu116/torch-1.12.0%2Bcu116-cp39-cp39-linux_x86_64.whl#sha256=7665e906995328746c6f70016ee90cafe50cbf434b6ef576e1de2678929ee63e
torch_scatter==2.1.2
torchmetrics==0.11.4
torchvision==0.13.1+cu116
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.28.1
trimesh==4.4.7
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.13
-e git+https://github.com/facebookresearch/xformers.git@1d31a3ac3b11f40fde7f00aa64debb0fd4d6f376#egg=xformers
yarl==1.9.4
zipp==3.19.2
